Modified program not only to be compiled by C99 but also compiled by CPP11
