+++ system/drivers/st/lsm303agr/LSM303AGR_ACC_driver.c
@@ -36,38 +36,41 @@
 /* Private variables ---------------------------------------------------------*/
 
 /* Private functions ---------------------------------------------------------*/
-
-
 /*******************************************************************************
-* Function Name		: SwapHighLowByte
-* Description		: Swap High/low byte in multiple byte values
-*                     It works with minimum 2 byte for every dimension.
-*                     Example x,y,z with 2 byte for every dimension
-*
-* Input				: bufferToSwap -> buffer to swap
-*                     numberOfByte -> the buffer length in byte
-*                     dimension -> number of dimension
-*
-* Output			: bufferToSwap -> buffer swapped
+* Function Name		: LSM303AGR_ACC_ReadReg
+* Description		: Generic Reading function. It must be fullfilled with either
+*					: I2C or SPI reading functions
+* Input				: Register Address
+* Output			: Data REad
 * Return			: None
 *******************************************************************************/
-void LSM303AGR_ACC_SwapHighLowByte(u8_t *bufferToSwap, u8_t numberOfByte, u8_t dimension)
+u8_t LSM303AGR_ACC_ReadReg(u8_t Reg, u8_t* Data)
 {
 
-    u8_t numberOfByteForDimension, i, j;
-    u8_t tempValue[10];
-
-    numberOfByteForDimension=numberOfByte/dimension;
-
-    for (i=0; i<dimension; i++ )
-    {
-        for (j=0; j<numberOfByteForDimension; j++ )
-            tempValue[j]=bufferToSwap[j+i*numberOfByteForDimension];
-        for (j=0; j<numberOfByteForDimension; j++ )
-            *(bufferToSwap+i*(numberOfByteForDimension)+j)=*(tempValue+(numberOfByteForDimension-1)-j);
-    }
+    //To be completed with either I2c or SPI reading function
+    //i.e.: *Data = SPI_Mems_Read_Reg( Reg );
+    if(!I2C_EXPBD_ReadData(Data, LSM303AGR_ACC_I2C_ADDRESS, Reg, 1))  //[Example]
+        return MEMS_ERROR;                                                        //[Example]
+    else                                                                        //[Example]
+        return MEMS_SUCCESS;                                                      //[Example]
 }
 
+/*******************************************************************************
+* Function Name		: LSM303AGR_ACC_WriteReg
+* Description		: Generic Writing function. It must be fullfilled with either
+*					: I2C or SPI writing function
+* Input				: Register Address, Data to be written
+* Output			: None
+* Return			: None
+*******************************************************************************/
+u8_t LSM303AGR_ACC_WriteReg(u8_t Reg, u8_t Data)
+{
+
+    //To be completed with either I2c or SPI writing function
+    //i.e.: SPI_Mems_Write_Reg(Reg, Data);
+    I2C_EXPBD_WriteData(&Data,  LSM303AGR_ACC_I2C_ADDRESS,  Reg, 1);       //[Example]
+    return MEMS_SUCCESS;                                                        //[Example]
+}
 /* Exported functions ---------------------------------------------------------*/
 
 /*******************************************************************************
@@ -83,8 +86,9 @@ status_t LSM303AGR_ACC_R_x_data_avail(LSM303AGR_ACC_1DA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_1DA_MASK; //mask
-
+    //*value &= LSM303AGR_ACC_1DA_MASK; //mask
+		*value = (LSM303AGR_ACC_1DA_t)((*value) & LSM303AGR_ACC_1DA_MASK);
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -100,8 +104,9 @@ status_t LSM303AGR_ACC_R_y_data_avail(LSM303AGR_ACC_2DA__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_2DA__MASK; //mask
-
+    //*value &= LSM303AGR_ACC_2DA__MASK; //mask
+    *value = (LSM303AGR_ACC_2DA__t)((*value) & LSM303AGR_ACC_2DA__MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -117,8 +122,9 @@ status_t LSM303AGR_ACC_R_z_data_avail(LSM303AGR_ACC_3DA__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_3DA__MASK; //mask
-
+    //*value &= LSM303AGR_ACC_3DA__MASK; //mask
+    *value = (LSM303AGR_ACC_3DA__t)((*value) & LSM303AGR_ACC_3DA__MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -134,8 +140,9 @@ status_t LSM303AGR_ACC_R_xyz_data_avail(LSM303AGR_ACC_321DA__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_321DA__MASK; //mask
-
+    //*value &= LSM303AGR_ACC_321DA__MASK; //mask
+    *value = (LSM303AGR_ACC_321DA__t)((*value) & LSM303AGR_ACC_321DA__MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -151,8 +158,9 @@ status_t LSM303AGR_ACC_R_DataXOverrun(LSM303AGR_ACC_1OR__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_1OR__MASK; //mask
-
+    //*value &= LSM303AGR_ACC_1OR__MASK; //mask
+    *value = (LSM303AGR_ACC_1OR__t)((*value) & LSM303AGR_ACC_1OR__MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -168,7 +176,8 @@ status_t LSM303AGR_ACC_R_DataYOverrun(LSM303AGR_ACC_2OR__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_2OR__MASK; //mask
+    //*value &= LSM303AGR_ACC_2OR__MASK; //mask
+    *value = (LSM303AGR_ACC_2OR__t)((*value) & LSM303AGR_ACC_2OR__MASK); //mask		
 
     return MEMS_SUCCESS;
 }
@@ -185,8 +194,9 @@ status_t LSM303AGR_ACC_R_DataZOverrun(LSM303AGR_ACC_3OR__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_3OR__MASK; //mask
-
+    //*value &= LSM303AGR_ACC_3OR__MASK; //mask
+    *value = (LSM303AGR_ACC_3OR__t)((*value) & LSM303AGR_ACC_3OR__MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -202,8 +212,9 @@ status_t LSM303AGR_ACC_R_DataXYZOverrun(LSM303AGR_ACC_321OR__t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG_AUX, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_321OR__MASK; //mask
-
+    //*value &= LSM303AGR_ACC_321OR__MASK; //mask
+    *value = (LSM303AGR_ACC_321OR__t)((*value) & LSM303AGR_ACC_321OR__MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -278,8 +289,9 @@ status_t LSM303AGR_ACC_R_TEMP_EN_bits(LSM303AGR_ACC_TEMP_EN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_TEMP_CFG_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_TEMP_EN_MASK; //mask
-
+    //*value &= LSM303AGR_ACC_TEMP_EN_MASK; //mask
+    *value = (LSM303AGR_ACC_TEMP_EN_t)((*value) & LSM303AGR_ACC_TEMP_EN_MASK); //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -318,8 +330,9 @@ status_t LSM303AGR_ACC_R_ADC_PD(LSM303AGR_ACC_ADC_PD_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_TEMP_CFG_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ADC_PD_MASK; //mask
-
+    //*value &= LSM303AGR_ACC_ADC_PD_MASK; //mask
+    *value = (LSM303AGR_ACC_ADC_PD_t)((*value) & LSM303AGR_ACC_ADC_PD_MASK);
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -358,8 +371,9 @@ status_t LSM303AGR_ACC_R_XEN(LSM303AGR_ACC_XEN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG1, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XEN_MASK; //mask
-
+    //*value &= LSM303AGR_ACC_XEN_MASK; //mask
+    *value = (LSM303AGR_ACC_XEN_t)LSM303AGR_ACC_XEN_MASK; //mask
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -398,8 +412,9 @@ status_t LSM303AGR_ACC_R_YEN(LSM303AGR_ACC_YEN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG1, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YEN_MASK; //mask
-
+    //*value &= LSM303AGR_ACC_YEN_MASK; //mask
+    *value = (LSM303AGR_ACC_YEN_t)((*value) & LSM303AGR_ACC_YEN_MASK);
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -438,7 +453,8 @@ status_t LSM303AGR_ACC_R_ZEN(LSM303AGR_ACC_ZEN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG1, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZEN_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZEN_MASK; //mask
+		*value = (LSM303AGR_ACC_ZEN_t)((*value) & LSM303AGR_ACC_ZEN_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -478,7 +494,8 @@ status_t LSM303AGR_ACC_R_LOWPWR_EN(LSM303AGR_ACC_LPEN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG1, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_LPEN_MASK; //mask
+    //*value &= LSM303AGR_ACC_LPEN_MASK; //mask
+		*value = (LSM303AGR_ACC_LPEN_t)((*value) & LSM303AGR_ACC_LPEN_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -518,7 +535,8 @@ status_t LSM303AGR_ACC_R_ODR(LSM303AGR_ACC_ODR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG1, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ODR_MASK; //mask
+    //*value &= LSM303AGR_ACC_ODR_MASK; //mask
+		*value = (LSM303AGR_ACC_ODR_t)((*value) & LSM303AGR_ACC_ODR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -558,7 +576,8 @@ status_t LSM303AGR_ACC_R_hpf_aoi_en_int1(LSM303AGR_ACC_HPIS1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_HPIS1_MASK; //mask
+    //*value &= LSM303AGR_ACC_HPIS1_MASK; //mask
+		*value = (LSM303AGR_ACC_HPIS1_t)((*value) & LSM303AGR_ACC_HPIS1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -598,7 +617,8 @@ status_t LSM303AGR_ACC_R_hpf_aoi_en_int2(LSM303AGR_ACC_HPIS2_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_HPIS2_MASK; //mask
+    //*value &= LSM303AGR_ACC_HPIS2_MASK; //mask
+		*value = (LSM303AGR_ACC_HPIS2_t)((*value) & LSM303AGR_ACC_HPIS2_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -638,7 +658,8 @@ status_t LSM303AGR_ACC_R_hpf_click_en(LSM303AGR_ACC_HPCLICK_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_HPCLICK_MASK; //mask
+    //*value &= LSM303AGR_ACC_HPCLICK_MASK; //mask
+		*value = (LSM303AGR_ACC_HPCLICK_t)(*value & LSM303AGR_ACC_HPCLICK_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -678,7 +699,8 @@ status_t LSM303AGR_ACC_R_Data_Filter(LSM303AGR_ACC_FDS_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_FDS_MASK; //mask
+    //*value &= LSM303AGR_ACC_FDS_MASK; //mask
+		*value = (LSM303AGR_ACC_FDS_t)(*value & LSM303AGR_ACC_FDS_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -718,7 +740,8 @@ status_t LSM303AGR_ACC_R_hpf_cutoff_freq(LSM303AGR_ACC_HPCF_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_HPCF_MASK; //mask
+    //*value &= LSM303AGR_ACC_HPCF_MASK; //mask
+		*value = (LSM303AGR_ACC_HPCF_t)((*value) & LSM303AGR_ACC_HPCF_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -758,7 +781,8 @@ status_t LSM303AGR_ACC_R_hpf_mode(LSM303AGR_ACC_HPM_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_HPM_MASK; //mask
+    //*value &= LSM303AGR_ACC_HPM_MASK; //mask
+		*value = (LSM303AGR_ACC_HPM_t)((*value) & LSM303AGR_ACC_HPM_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -798,7 +822,8 @@ status_t LSM303AGR_ACC_R_FIFO_Overrun_on_INT1(LSM303AGR_ACC_I1_OVERRUN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_OVERRUN_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_OVERRUN_MASK; //mask
+    *value = (LSM303AGR_ACC_I1_OVERRUN_t)((*value) & LSM303AGR_ACC_I1_OVERRUN_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -838,7 +863,8 @@ status_t LSM303AGR_ACC_R_FIFO_Watermark_on_INT1(LSM303AGR_ACC_I1_WTM_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_WTM_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_WTM_MASK; //mask
+    *value = (LSM303AGR_ACC_I1_WTM_t)((*value) & LSM303AGR_ACC_I1_WTM_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -878,7 +904,8 @@ status_t LSM303AGR_ACC_R_FIFO_DRDY2_on_INT1(LSM303AGR_ACC_I1_DRDY2_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_DRDY2_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_DRDY2_MASK; //mask		
+    *value = (LSM303AGR_ACC_I1_DRDY2_t)((*value) & LSM303AGR_ACC_I1_DRDY2_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -918,7 +945,8 @@ status_t LSM303AGR_ACC_R_FIFO_DRDY1_on_INT1(LSM303AGR_ACC_I1_DRDY1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_DRDY1_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_DRDY1_MASK; //mask		
+    *value = (LSM303AGR_ACC_I1_DRDY1_t)((*value) & LSM303AGR_ACC_I1_DRDY1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -958,7 +986,8 @@ status_t LSM303AGR_ACC_R_FIFO_AOL2_on_INT1(LSM303AGR_ACC_I1_AOI2_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_AOI2_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_AOI2_MASK; //mask		
+    *value = (LSM303AGR_ACC_I1_AOI2_t)((*value) & LSM303AGR_ACC_I1_AOI2_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -998,7 +1027,8 @@ status_t LSM303AGR_ACC_R_FIFO_AOL1_on_INT1(LSM303AGR_ACC_I1_AOI1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_AOI1_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_AOI1_MASK; //mask		
+    *value = (LSM303AGR_ACC_I1_AOI1_t)((*value) & LSM303AGR_ACC_I1_AOI1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1038,7 +1068,8 @@ status_t LSM303AGR_ACC_R_FIFO_Click_on_INT1(LSM303AGR_ACC_I1_CLICK_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG3, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I1_CLICK_MASK; //mask
+    //*value &= LSM303AGR_ACC_I1_CLICK_MASK; //mask		
+    *value = (LSM303AGR_ACC_I1_CLICK_t)((*value) & LSM303AGR_ACC_I1_CLICK_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1078,8 +1109,9 @@ status_t LSM303AGR_ACC_R_SPI_mode(LSM303AGR_ACC_SIM_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG4, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_SIM_MASK; //mask
-
+    //*value &= LSM303AGR_ACC_SIM_MASK; //mask
+    *value = (LSM303AGR_ACC_SIM_t)((*value) & LSM303AGR_ACC_SIM_MASK);
+		
     return MEMS_SUCCESS;
 }
 /*******************************************************************************
@@ -1118,7 +1150,8 @@ status_t LSM303AGR_ACC_R_SelfTest(LSM303AGR_ACC_ST_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG4, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ST_MASK; //mask
+    //*value &= LSM303AGR_ACC_ST_MASK; //mask
+    *value = (LSM303AGR_ACC_ST_t)((*value) & LSM303AGR_ACC_ST_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1158,7 +1191,8 @@ status_t LSM303AGR_ACC_R_HiRes(LSM303AGR_ACC_HR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG4, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_HR_MASK; //mask
+    //*value &= LSM303AGR_ACC_HR_MASK; //mask
+    *value = (LSM303AGR_ACC_HR_t)((*value) & LSM303AGR_ACC_HR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1198,7 +1232,8 @@ status_t LSM303AGR_ACC_R_FullScale(LSM303AGR_ACC_FS_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG4, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_FS_MASK; //mask
+    //*value &= LSM303AGR_ACC_FS_MASK; //mask
+    *value = (LSM303AGR_ACC_FS_t)((*value) & LSM303AGR_ACC_FS_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1238,7 +1273,8 @@ status_t LSM303AGR_ACC_R_LittleBigEndian(LSM303AGR_ACC_BLE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG4, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_BLE_MASK; //mask
+    //*value &= LSM303AGR_ACC_BLE_MASK; //mask
+    *value = (LSM303AGR_ACC_BLE_t)((*value) & LSM303AGR_ACC_BLE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1278,7 +1314,8 @@ status_t LSM303AGR_ACC_R_BlockDataUpdate(LSM303AGR_ACC_BDU_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG4, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_BDU_MASK; //mask
+    //*value &= LSM303AGR_ACC_BDU_MASK; //mask
+    *value = (LSM303AGR_ACC_BDU_t)((*value) & LSM303AGR_ACC_BDU_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1318,7 +1355,8 @@ status_t LSM303AGR_ACC_R_4D_on_INT2(LSM303AGR_ACC_D4D_INT2_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG5, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_D4D_INT2_MASK; //mask
+    //*value &= LSM303AGR_ACC_D4D_INT2_MASK; //mask
+    *value = (LSM303AGR_ACC_D4D_INT2_t)((*value) & LSM303AGR_ACC_D4D_INT2_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1358,7 +1396,8 @@ status_t LSM303AGR_ACC_R_LatchInterrupt_on_INT2(LSM303AGR_ACC_LIR_INT2_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG5, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_LIR_INT2_MASK; //mask
+    //*value &= LSM303AGR_ACC_LIR_INT2_MASK; //mask
+    *value = (LSM303AGR_ACC_LIR_INT2_t)((*value) & LSM303AGR_ACC_LIR_INT2_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1399,7 +1438,8 @@ status_t LSM303AGR_ACC_R_4D_on_INT1(LSM303AGR_ACC_D4D_INT1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG5, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_D4D_INT1_MASK; //mask
+    //*value &= LSM303AGR_ACC_D4D_INT1_MASK; //mask
+    *value = (LSM303AGR_ACC_D4D_INT1_t)((*value) & LSM303AGR_ACC_D4D_INT1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1439,7 +1479,8 @@ status_t LSM303AGR_ACC_R_LatchInterrupt_on_INT1(LSM303AGR_ACC_LIR_INT1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG5, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_LIR_INT1_MASK; //mask
+    //*value &= LSM303AGR_ACC_LIR_INT1_MASK; //mask
+    *value = (LSM303AGR_ACC_LIR_INT1_t)((*value) & LSM303AGR_ACC_LIR_INT1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1480,7 +1521,8 @@ status_t LSM303AGR_ACC_R_FIFO_EN(LSM303AGR_ACC_FIFO_EN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG5, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_FIFO_EN_MASK; //mask
+    //*value &= LSM303AGR_ACC_FIFO_EN_MASK; //mask
+    *value = (LSM303AGR_ACC_FIFO_EN_t)((*value) & LSM303AGR_ACC_FIFO_EN_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1520,7 +1562,8 @@ status_t LSM303AGR_ACC_R_RebootMemory(LSM303AGR_ACC_BOOT_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG5, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_BOOT_MASK; //mask
+    //*value &= LSM303AGR_ACC_BOOT_MASK; //mask
+    *value = (LSM303AGR_ACC_BOOT_t)((*value) & LSM303AGR_ACC_BOOT_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1560,7 +1603,8 @@ status_t LSM303AGR_ACC_R_IntActive(LSM303AGR_ACC_H_LACTIVE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG6, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_H_LACTIVE_MASK; //mask
+    //*value &= LSM303AGR_ACC_H_LACTIVE_MASK; //mask
+    *value = (LSM303AGR_ACC_H_LACTIVE_t)((*value) & LSM303AGR_ACC_H_LACTIVE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1601,7 +1645,8 @@ status_t LSM303AGR_ACC_R_P2_ACT(LSM303AGR_ACC_P2_ACT_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG6, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_P2_ACT_MASK; //mask
+    //*value &= LSM303AGR_ACC_P2_ACT_MASK; //mask
+    *value = (LSM303AGR_ACC_P2_ACT_t)((*value) & LSM303AGR_ACC_P2_ACT_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1642,7 +1687,8 @@ status_t LSM303AGR_ACC_R_Boot_on_INT2(LSM303AGR_ACC_BOOT_I1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG6, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_BOOT_I1_MASK; //mask
+    //*value &= LSM303AGR_ACC_BOOT_I1_MASK; //mask
+    *value = (LSM303AGR_ACC_BOOT_I1_t)((*value) & LSM303AGR_ACC_BOOT_I1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1682,7 +1728,8 @@ status_t LSM303AGR_ACC_R_I2_on_INT2(LSM303AGR_ACC_I2_INT2_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG6, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I2_INT2_MASK; //mask
+    //*value &= LSM303AGR_ACC_I2_INT2_MASK; //mask
+    *value = (LSM303AGR_ACC_I2_INT2_t)((*value) & LSM303AGR_ACC_I2_INT2_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1722,7 +1769,8 @@ status_t LSM303AGR_ACC_R_I2_on_INT1(LSM303AGR_ACC_I2_INT1_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG6, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I2_INT1_MASK; //mask
+    //*value &= LSM303AGR_ACC_I2_INT1_MASK; //mask
+    *value = (LSM303AGR_ACC_I2_INT1_t)((*value) & LSM303AGR_ACC_I2_INT1_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1762,7 +1810,8 @@ status_t LSM303AGR_ACC_R_Click_on_INT2(LSM303AGR_ACC_I2_CLICKEN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CTRL_REG6, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_I2_CLICKEN_MASK; //mask
+    //*value &= LSM303AGR_ACC_I2_CLICKEN_MASK; //mask
+    *value = (LSM303AGR_ACC_I2_CLICKEN_t)((*value) & LSM303AGR_ACC_I2_CLICKEN_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1823,7 +1872,8 @@ status_t LSM303AGR_ACC_R_XDataAvail(LSM303AGR_ACC_XDA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XDA_MASK; //mask
+    //*value &= LSM303AGR_ACC_XDA_MASK; //mask
+    *value = (LSM303AGR_ACC_XDA_t)((*value) & LSM303AGR_ACC_XDA_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1840,7 +1890,8 @@ status_t LSM303AGR_ACC_R_YDataAvail(LSM303AGR_ACC_YDA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YDA_MASK; //mask
+    //*value &= LSM303AGR_ACC_YDA_MASK; //mask
+    *value = (LSM303AGR_ACC_YDA_t)((*value) & LSM303AGR_ACC_YDA_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1857,7 +1908,8 @@ status_t LSM303AGR_ACC_R_ZDataAvail(LSM303AGR_ACC_ZDA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZDA_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZDA_MASK; //mask
+    *value = (LSM303AGR_ACC_ZDA_t)((*value) & LSM303AGR_ACC_ZDA_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1874,7 +1926,8 @@ status_t LSM303AGR_ACC_R_XYZDataAvail(LSM303AGR_ACC_ZYXDA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZYXDA_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZYXDA_MASK; //mask
+    *value = (LSM303AGR_ACC_ZYXDA_t)((*value) & LSM303AGR_ACC_ZYXDA_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1891,7 +1944,8 @@ status_t LSM303AGR_ACC_R_XDataOverrun(LSM303AGR_ACC_XOR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XOR_MASK; //mask
+    //*value &= LSM303AGR_ACC_XOR_MASK; //mask
+    *value = (LSM303AGR_ACC_XOR_t)((*value) & LSM303AGR_ACC_XOR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1908,7 +1962,8 @@ status_t LSM303AGR_ACC_R_YDataOverrun(LSM303AGR_ACC_YOR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YOR_MASK; //mask
+    //*value &= LSM303AGR_ACC_YOR_MASK; //mask
+    *value = (LSM303AGR_ACC_YOR_t)((*value) & LSM303AGR_ACC_YOR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1925,7 +1980,8 @@ status_t LSM303AGR_ACC_R_ZDataOverrun(LSM303AGR_ACC_ZOR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZOR_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZOR_MASK; //mask
+    *value = (LSM303AGR_ACC_ZOR_t)((*value) & LSM303AGR_ACC_ZOR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -1942,7 +1998,8 @@ status_t LSM303AGR_ACC_R_XYZDataOverrun(LSM303AGR_ACC_ZYXOR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_STATUS_REG2, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZYXOR_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZYXOR_MASK; //mask
+    *value = (LSM303AGR_ACC_ZYXOR_t)((*value) & LSM303AGR_ACC_ZYXOR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2026,7 +2083,8 @@ status_t LSM303AGR_ACC_R_TriggerSel(LSM303AGR_ACC_TR_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_FIFO_CTRL_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_TR_MASK; //mask
+    //*value &= LSM303AGR_ACC_TR_MASK; //mask
+		*value = (LSM303AGR_ACC_TR_t)((*value) & LSM303AGR_ACC_TR_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2066,7 +2124,8 @@ status_t LSM303AGR_ACC_R_FifoMode(LSM303AGR_ACC_FM_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_FIFO_CTRL_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_FM_MASK; //mask
+    //*value &= LSM303AGR_ACC_FM_MASK; //mask
+    *value = (LSM303AGR_ACC_FM_t)((*value) & LSM303AGR_ACC_FM_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2101,7 +2160,9 @@ status_t LSM303AGR_ACC_R_FifoEmpty(LSM303AGR_ACC_EMPTY_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_FIFO_SRC_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_EMPTY_MASK; //mask
+    //*value &= LSM303AGR_ACC_EMPTY_MASK; //mask
+    *value = (LSM303AGR_ACC_EMPTY_t)((*value) & LSM303AGR_ACC_EMPTY_MASK);
+    *value = (LSM303AGR_ACC_EMPTY_t)(*value >> LSM303AGR_ACC_FSS_POSITION); //mask
 
     return MEMS_SUCCESS;
 }
@@ -2118,7 +2179,8 @@ status_t LSM303AGR_ACC_R_FifoOverrun(LSM303AGR_ACC_OVRN_FIFO_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_FIFO_SRC_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_OVRN_FIFO_MASK; //mask
+    //*value &= LSM303AGR_ACC_OVRN_FIFO_MASK; //mask
+    *value = (LSM303AGR_ACC_OVRN_FIFO_t)((*value) & LSM303AGR_ACC_OVRN_FIFO_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2135,7 +2197,8 @@ status_t LSM303AGR_ACC_R_WatermarkLevel(LSM303AGR_ACC_WTM_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_FIFO_SRC_REG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_WTM_MASK; //mask
+    //*value &= LSM303AGR_ACC_WTM_MASK; //mask
+    *value = (LSM303AGR_ACC_WTM_t)((*value) & LSM303AGR_ACC_WTM_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2175,7 +2238,8 @@ status_t LSM303AGR_ACC_R_Int1EnXLo(LSM303AGR_ACC_XLIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XLIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_XLIE_MASK; //mask
+    *value = (LSM303AGR_ACC_XLIE_t)((*value) & LSM303AGR_ACC_XLIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2215,7 +2279,8 @@ status_t LSM303AGR_ACC_R_Int1EnXHi(LSM303AGR_ACC_XHIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XHIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_XHIE_MASK; //mask
+    *value = (LSM303AGR_ACC_XHIE_t)((*value) & LSM303AGR_ACC_XHIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2255,7 +2320,8 @@ status_t LSM303AGR_ACC_R_Int1EnYLo(LSM303AGR_ACC_YLIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YLIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_YLIE_MASK; //mask
+    *value = (LSM303AGR_ACC_YLIE_t)((*value) & LSM303AGR_ACC_YLIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2295,7 +2361,8 @@ status_t LSM303AGR_ACC_R_Int1EnYHi(LSM303AGR_ACC_YHIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YHIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_YHIE_MASK; //mask
+    *value = (LSM303AGR_ACC_YHIE_t)((*value) & LSM303AGR_ACC_YHIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2335,7 +2402,8 @@ status_t LSM303AGR_ACC_R_Int1EnZLo(LSM303AGR_ACC_ZLIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZLIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZLIE_MASK; //mask
+    *value = (LSM303AGR_ACC_ZLIE_t)((*value) & LSM303AGR_ACC_ZLIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2375,7 +2443,8 @@ status_t LSM303AGR_ACC_R_Int1EnZHi(LSM303AGR_ACC_ZHIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZHIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZHIE_MASK; //mask
+    *value = (LSM303AGR_ACC_ZHIE_t)((*value) & LSM303AGR_ACC_ZHIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2415,7 +2484,8 @@ status_t LSM303AGR_ACC_R_Int1_6D(LSM303AGR_ACC_6D_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_6D_MASK; //mask
+    //*value &= LSM303AGR_ACC_6D_MASK; //mask
+    *value = (LSM303AGR_ACC_6D_t)((*value) & LSM303AGR_ACC_6D_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2455,7 +2525,8 @@ status_t LSM303AGR_ACC_R_Int1_AOI(LSM303AGR_ACC_AOI_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_AOI_MASK; //mask
+    //*value &= LSM303AGR_ACC_AOI_MASK; //mask
+    *value = (LSM303AGR_ACC_AOI_t)((*value) & LSM303AGR_ACC_AOI_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2495,7 +2566,8 @@ status_t LSM303AGR_ACC_R_Int2EnXLo(LSM303AGR_ACC_XLIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XLIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_XLIE_MASK; //mask
+    *value = (LSM303AGR_ACC_XLIE_t)((*value) & LSM303AGR_ACC_XLIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2535,7 +2607,8 @@ status_t LSM303AGR_ACC_R_Int2EnXHi(LSM303AGR_ACC_XHIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XHIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_XHIE_MASK; //mask
+    *value = (LSM303AGR_ACC_XHIE_t)((*value) & LSM303AGR_ACC_XHIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2575,7 +2648,8 @@ status_t LSM303AGR_ACC_R_Int2EnYLo(LSM303AGR_ACC_YLIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YLIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_YLIE_MASK; //mask
+    *value = (LSM303AGR_ACC_YLIE_t)((*value) & LSM303AGR_ACC_YLIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2615,7 +2689,8 @@ status_t LSM303AGR_ACC_R_Int2EnYHi(LSM303AGR_ACC_YHIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YHIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_YHIE_MASK; //mask
+		*value = (LSM303AGR_ACC_YHIE_t)((*value) & LSM303AGR_ACC_YHIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2655,7 +2730,8 @@ status_t LSM303AGR_ACC_R_Int2EnZLo(LSM303AGR_ACC_ZLIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZLIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZLIE_MASK; //mask
+		*value = (LSM303AGR_ACC_ZLIE_t)((*value) & LSM303AGR_ACC_ZLIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2695,7 +2771,8 @@ status_t LSM303AGR_ACC_R_Int2EnZHi(LSM303AGR_ACC_ZHIE_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZHIE_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZHIE_MASK; //mask
+		*value = (LSM303AGR_ACC_ZHIE_t)((*value) & LSM303AGR_ACC_ZHIE_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2735,7 +2812,8 @@ status_t LSM303AGR_ACC_R_Int2_6D(LSM303AGR_ACC_6D_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_6D_MASK; //mask
+    //*value &= LSM303AGR_ACC_6D_MASK; //mask
+		*value = (LSM303AGR_ACC_6D_t)((*value) & LSM303AGR_ACC_6D_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2775,7 +2853,8 @@ status_t LSM303AGR_ACC_R_Int2_AOI(LSM303AGR_ACC_AOI_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_AOI_MASK; //mask
+    //*value &= LSM303AGR_ACC_AOI_MASK; //mask
+		*value = (LSM303AGR_ACC_AOI_t)((*value) & LSM303AGR_ACC_AOI_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2793,7 +2872,8 @@ status_t LSM303AGR_ACC_R_Int1_Xlo(LSM303AGR_ACC_XL_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XL_MASK; //mask
+    //*value &= LSM303AGR_ACC_XL_MASK; //mask
+		*value = (LSM303AGR_ACC_XL_t)((*value) & LSM303AGR_ACC_XL_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2810,7 +2890,8 @@ status_t LSM303AGR_ACC_R_Int1_XHi(LSM303AGR_ACC_XH_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XH_MASK; //mask
+    //*value &= LSM303AGR_ACC_XH_MASK; //mask
+		*value = (LSM303AGR_ACC_XH_t)((*value) & LSM303AGR_ACC_XH_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2827,7 +2908,8 @@ status_t LSM303AGR_ACC_R_Int1_YLo(LSM303AGR_ACC_YL_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YL_MASK; //mask
+    //*value &= LSM303AGR_ACC_YL_MASK; //mask
+		*value = (LSM303AGR_ACC_YL_t)((*value) & LSM303AGR_ACC_YL_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2844,7 +2926,8 @@ status_t LSM303AGR_ACC_R_Int1_YHi(LSM303AGR_ACC_YH_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YH_MASK; //mask
+    //*value &= LSM303AGR_ACC_YH_MASK; //mask
+		*value = (LSM303AGR_ACC_YH_t)((*value) & LSM303AGR_ACC_YH_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2861,7 +2944,8 @@ status_t LSM303AGR_ACC_R_Int1_Zlo(LSM303AGR_ACC_ZL_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZL_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZL_MASK; //mask
+		*value = (LSM303AGR_ACC_ZL_t)((*value) & LSM303AGR_ACC_ZL_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2878,7 +2962,8 @@ status_t LSM303AGR_ACC_R_Int1_ZHi(LSM303AGR_ACC_ZH_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZH_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZH_MASK; //mask
+		*value = (LSM303AGR_ACC_ZH_t)((*value) & LSM303AGR_ACC_ZH_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2895,7 +2980,8 @@ status_t LSM303AGR_ACC_R_Int1_IA(LSM303AGR_ACC_IA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT1_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_IA_MASK; //mask
+    //*value &= LSM303AGR_ACC_IA_MASK; //mask
+		*value = (LSM303AGR_ACC_IA_t)((*value) & LSM303AGR_ACC_IA_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2913,7 +2999,8 @@ status_t LSM303AGR_ACC_R_Int2_Xlo(LSM303AGR_ACC_XL_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XL_MASK; //mask
+    //*value &= LSM303AGR_ACC_XL_MASK; //mask
+		*value = (LSM303AGR_ACC_XL_t)((*value) & LSM303AGR_ACC_XL_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2930,7 +3017,8 @@ status_t LSM303AGR_ACC_R_Int2_XHi(LSM303AGR_ACC_XH_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XH_MASK; //mask
+    //*value &= LSM303AGR_ACC_XH_MASK; //mask
+		*value = (LSM303AGR_ACC_XH_t)((*value) & LSM303AGR_ACC_XH_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2947,7 +3035,8 @@ status_t LSM303AGR_ACC_R_Int2_YLo(LSM303AGR_ACC_YL_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YL_MASK; //mask
+    //*value &= LSM303AGR_ACC_YL_MASK; //mask
+		*value = (LSM303AGR_ACC_YL_t)((*value) & LSM303AGR_ACC_YL_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2964,7 +3053,8 @@ status_t LSM303AGR_ACC_R_Int2_YHi(LSM303AGR_ACC_YH_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YH_MASK; //mask
+    //*value &= LSM303AGR_ACC_YH_MASK; //mask
+		*value = (LSM303AGR_ACC_YH_t)((*value) & LSM303AGR_ACC_YH_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2981,7 +3071,8 @@ status_t LSM303AGR_ACC_R_Int2_Zlo(LSM303AGR_ACC_ZL_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZL_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZL_MASK; //mask
+		*value = (LSM303AGR_ACC_ZL_t)((*value) & LSM303AGR_ACC_ZL_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -2998,7 +3089,8 @@ status_t LSM303AGR_ACC_R_Int2_ZHi(LSM303AGR_ACC_ZH_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZH_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZH_MASK; //mask
+		*value = (LSM303AGR_ACC_ZH_t)((*value) & LSM303AGR_ACC_ZH_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3015,7 +3107,8 @@ status_t LSM303AGR_ACC_R_Int2_IA(LSM303AGR_ACC_IA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_INT2_SOURCE, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_IA_MASK; //mask
+    //*value &= LSM303AGR_ACC_IA_MASK; //mask
+		*value = (LSM303AGR_ACC_IA_t)((*value) & LSM303AGR_ACC_IA_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3236,7 +3329,8 @@ status_t LSM303AGR_ACC_R_XSingle(LSM303AGR_ACC_XS_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XS_MASK; //mask
+    //*value &= LSM303AGR_ACC_XS_MASK; //mask
+		*value = (LSM303AGR_ACC_XS_t)((*value) & LSM303AGR_ACC_XS_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3276,7 +3370,8 @@ status_t LSM303AGR_ACC_R_XDouble(LSM303AGR_ACC_XD_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_XD_MASK; //mask
+    //*value &= LSM303AGR_ACC_XD_MASK; //mask
+		*value = (LSM303AGR_ACC_XD_t)((*value) & LSM303AGR_ACC_XD_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3316,7 +3411,8 @@ status_t LSM303AGR_ACC_R_YSingle(LSM303AGR_ACC_YS_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YS_MASK; //mask
+    //*value &= LSM303AGR_ACC_YS_MASK; //mask
+		*value = (LSM303AGR_ACC_YS_t)((*value) & LSM303AGR_ACC_YS_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3356,7 +3452,8 @@ status_t LSM303AGR_ACC_R_YDouble(LSM303AGR_ACC_YD_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_YD_MASK; //mask
+    //*value &= LSM303AGR_ACC_YD_MASK; //mask
+		*value = (LSM303AGR_ACC_YD_t)((*value) & LSM303AGR_ACC_YD_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3396,7 +3493,8 @@ status_t LSM303AGR_ACC_R_ZSingle(LSM303AGR_ACC_ZS_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZS_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZS_MASK; //mask
+		*value = (LSM303AGR_ACC_ZS_t)((*value) & LSM303AGR_ACC_ZS_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3436,7 +3534,8 @@ status_t LSM303AGR_ACC_R_ZDouble(LSM303AGR_ACC_ZD_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_CFG, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_ZD_MASK; //mask
+    //*value &= LSM303AGR_ACC_ZD_MASK; //mask
+		*value = (LSM303AGR_ACC_ZD_t)((*value) & LSM303AGR_ACC_ZD_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3453,7 +3552,8 @@ status_t LSM303AGR_ACC_R_ClickX(LSM303AGR_ACC_X_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_X_MASK; //mask
+    //*value &= LSM303AGR_ACC_X_MASK; //mask
+		*value = (LSM303AGR_ACC_X_t)((*value) & LSM303AGR_ACC_X_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3470,7 +3570,8 @@ status_t LSM303AGR_ACC_R_ClickY(LSM303AGR_ACC_Y_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_Y_MASK; //mask
+    //*value &= LSM303AGR_ACC_Y_MASK; //mask
+		*value = (LSM303AGR_ACC_Y_t)((*value) & LSM303AGR_ACC_Y_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3487,7 +3588,8 @@ status_t LSM303AGR_ACC_R_ClickZ(LSM303AGR_ACC_Z_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_Z_MASK; //mask
+    //*value &= LSM303AGR_ACC_Z_MASK; //mask
+		*value = (LSM303AGR_ACC_Z_t)((*value) & LSM303AGR_ACC_Z_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3504,7 +3606,8 @@ status_t LSM303AGR_ACC_R_ClickSign(LSM303AGR_ACC_SIGN_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_SIGN_MASK; //mask
+    //*value &= LSM303AGR_ACC_SIGN_MASK; //mask
+		*value = (LSM303AGR_ACC_SIGN_t)((*value) & LSM303AGR_ACC_SIGN_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3521,7 +3624,8 @@ status_t LSM303AGR_ACC_R_SingleCLICK(LSM303AGR_ACC_SCLICK_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_SCLICK_MASK; //mask
+    //*value &= LSM303AGR_ACC_SCLICK_MASK; //mask
+		*value = (LSM303AGR_ACC_SCLICK_t)((*value) & LSM303AGR_ACC_SCLICK_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3538,7 +3642,8 @@ status_t LSM303AGR_ACC_R_DoubleCLICK(LSM303AGR_ACC_DCLICK_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_DCLICK_MASK; //mask
+    //*value &= LSM303AGR_ACC_DCLICK_MASK; //mask
+		*value = (LSM303AGR_ACC_DCLICK_t)((*value) & LSM303AGR_ACC_DCLICK_MASK);
 
     return MEMS_SUCCESS;
 }
@@ -3555,7 +3660,8 @@ status_t LSM303AGR_ACC_R_CLICK_IA(LSM303AGR_ACC_CLICK_IA_t *value)
     if( HAL_OK!=LSM303AGR_ACC_ReadReg(LSM303AGR_ACC_CLICK_SRC, (u8_t *)value) )
         return MEMS_ERROR;
 
-    *value &= LSM303AGR_ACC_IA_MASK; //mask
+    //*value &= LSM303AGR_ACC_IA_MASK; //mask
+		*value = (LSM303AGR_ACC_CLICK_IA_t)((*value) & LSM303AGR_ACC_IA_MASK);
 
     return MEMS_SUCCESS;
 }
