+++ system/drivers/st/lsm6ds3/lsm6ds3.h
@@ -1439,9 +1439,7 @@ extern IMU_6AXES_StatusTypeDef LSM6DS3_IO_Read( uint8_t* pBuffer, uint8_t Device
 extern IMU_6AXES_StatusTypeDef LSM6DS3_IO_Read_DMA( uint8_t* pBuffer, uint8_t DeviceAddr, uint8_t RegisterAddr,
         uint16_t NumByteToRead );
 #endif
-extern void LSM6DS3_IO_ITConfig( void );
-/*magneto sensor IO functions*/
-extern void LSM303AGR_IO_ITConfig(void);
+
 
 /**
  * @}
